+++ src/plugins/clock/animation.cpp
@@ -12,26 +12,28 @@
 #include <debug_helper_disable.h>
 #endif
 
+using namespace Clock;
+
 // ------------------------------------------------------------------------
 // Color Class
 
-Clock::Color::Color() : _value(0)
+Color::Color() : _value(0)
 {
 }
 
-Clock::Color::Color(uint8_t *values) : _blue(values[0]), _green(values[1]), _red(values[2])
+Color::Color(uint8_t *values) : _blue(values[0]), _green(values[1]), _red(values[2])
 {
 }
 
-Clock::Color::Color(uint8_t red, uint8_t green, uint8_t blue) : _blue(blue), _green(green), _red(red)
+Color::Color(uint8_t red, uint8_t green, uint8_t blue) : _blue(blue), _green(green), _red(red)
 {
 }
 
-Clock::Color::Color(uint32_t value) : _value(value)
+Color::Color(uint32_t value) : _value(value)
 {
 }
 
-Clock::Color Clock::Color::fromString(const String &value)
+Color Color::fromString(const String &value)
 {
     auto ptr = value.c_str();
     while(isspace(*ptr) || *ptr == '#') {
@@ -40,12 +42,12 @@ Clock::Color Clock::Color::fromString(const String &value)
     return static_cast<uint32_t>(strtol(ptr, nullptr, 16) & 0xffffff);
 }
 
-Clock::Color Clock::Color::fromBGR(uint32_t value)
+Color Color::fromBGR(uint32_t value)
 {
     return Color(value, value >> 8, value >> 16);
 }
 
-String Clock::Color::toString() const
+String Color::toString() const
 {
     char buf[8];
     //snprintf_P(buf, sizeof(buf), PSTR("#%02X%02X%02X"), _red, _green, _blue);
@@ -53,14 +55,14 @@ String Clock::Color::toString() const
     return buf;
 }
 
-String Clock::Color::implode(char sep) const
+String Color::implode(char sep) const
 {
     char buf[16];
     snprintf_P(buf, sizeof(buf), PSTR("%u%c%u%c%u"), _red, sep, _green, sep, _blue);
     return buf;
 }
 
-uint8_t Clock::Color::_getRand(uint8_t mod, uint8_t mul, uint8_t factor)
+uint8_t Color::_getRand(uint8_t mod, uint8_t mul, uint8_t factor)
 {
     return rand() % mod * mul;
     // uint8_t rnd = rand() % mod;
@@ -70,7 +72,7 @@ uint8_t Clock::Color::_getRand(uint8_t mod, uint8_t mul, uint8_t factor)
     // return (rnd * mul * factor) / 255;
 }
 
-uint32_t Clock::Color::rnd(uint8_t minValue)
+uint32_t Color::rnd(uint8_t minValue)
 {
     do {
         _red = _getRand(kRndMod, kRndMul);
@@ -81,7 +83,7 @@ uint32_t Clock::Color::rnd(uint8_t minValue)
     return _value;
 }
 
-// uint32_t Clock::Color::rnd(Color factor, uint8_t minValue)
+// uint32_t Color::rnd(Color factor, uint8_t minValue)
 // {
 //     uint8_t minRed = red() * minValue / 255;
 //     uint8_t count = 0;
@@ -111,78 +113,78 @@ uint32_t Clock::Color::rnd(uint8_t minValue)
 //     return _value;
 // }
 
-uint32_t Clock::Color::get() const
+uint32_t Color::get() const
 {
     return _value;
 }
 
-Clock::Color &Clock::Color::operator=(uint32_t value)
+Color &Color::operator=(uint32_t value)
 {
     _value = value;
     return *this;
 }
 
-Clock::Color::operator bool() const
+Color::operator bool() const
 {
     return _value != 0;
 }
 
-Clock::Color::operator int() const
+Color::operator int() const
 {
     return _value;
 }
 
-Clock::Color::operator uint32_t() const
+Color::operator uint32_t() const
 {
     return _value;
 }
 
-bool Clock::Color::operator==(uint32_t value) const
+bool Color::operator==(uint32_t value) const
 {
     return _value == value;
 }
 
-bool Clock::Color::operator!=(uint32_t value) const
+bool Color::operator!=(uint32_t value) const
 {
     return _value != value;
 }
 
-bool Clock::Color::operator==(int value) const
+bool Color::operator==(int value) const
 {
     return _value == value;
 }
 
-bool Clock::Color::operator!=(int value) const
+bool Color::operator!=(int value) const
 {
     return _value != value;
 }
 
-uint8_t Clock::Color::red() const
+uint8_t Color::red() const
 {
     return _red;
 }
 
-uint8_t Clock::Color::green() const
+uint8_t Color::green() const
 {
     return _green;
 }
 
-uint8_t Clock::Color::blue() const
+uint8_t Color::blue() const
 {
     return _blue;
 }
 
-uint8_t &Clock::Color::red()
+uint8_t &Color::red()
 {
     return _red;
 }
 
-uint8_t &Clock::Color::green()
+uint8_t &Color::green()
 {
     return _green;
 }
 
-uint8_t &Clock::Color::blue()
+uint8_t &Color::blue()
 {
     return _blue;
 }
@@ -190,11 +192,11 @@ uint8_t &Clock::Color::blue()
 // ------------------------------------------------------------------------
 // Base Class Animation
 
-Clock::Animation::Animation(ClockPlugin &clock) : _clock(clock), _updateRate(_clock.getUpdateRate()), _finished(false), _blinkColon(true)
+Animation::Animation(ClockPlugin &clock) : _clock(clock), _updateRate(_clock.getUpdateRate()), _finished(false), _blinkColon(true)
 {
 }
 
-Clock::Animation::~Animation()
+Animation::~Animation()
 {
     if (!_finished) {
         end();
@@ -203,12 +205,12 @@ Clock::Animation::~Animation()
 }
 
 
-void Clock::Animation::begin()
+void Animation::begin()
 {
     __LDBG_printf("begin");
     _finished = true;
 }
-void Clock::Animation::end()
+void Animation::end()
 {
     __LDBG_printf("end");
     _finished = true;
@@ -217,67 +219,67 @@ void Clock::Animation::end()
     removeAnimationCallback();
 }
 
-bool Clock::Animation::finished() const
+bool Animation::finished() const
 {
     return _finished;
 }
 
-bool Clock::Animation::doBlinkColon() const
+bool Animation::doBlinkColon() const
 {
     return _blinkColon;
 }
 
-uint16_t Clock::Animation::getUpdateRate() const
+uint16_t Animation::getUpdateRate() const
 {
     return _updateRate;
 }
 
-void Clock::Animation::loop(time_t now)
+void Animation::loop(time_t now)
 {
     if (_callback) {
         _callback(now);
     }
 }
 
-void Clock::Animation::setUpdateRate(uint16_t updateRate)
+void Animation::setUpdateRate(uint16_t updateRate)
 {
     _updateRate = updateRate;
     _clock.setUpdateRate(_updateRate);
 }
 
-uint16_t Clock::Animation::getClockUpdateRate() const
+uint16_t Animation::getClockUpdateRate() const
 {
     return _clock.getUpdateRate();
 }
 
 
-void Clock::Animation::setColor(Color color)
+void Animation::setColor(Color color)
 {
     _clock.setColor(color);
 }
 
-Clock::Color Clock::Animation::getColor() const
+Color Animation::getColor() const
 {
     return _clock.getColor();
 }
 
 
-void Clock::Animation::setAnimationCallback(AnimationCallback callback)
+void Animation::setAnimationCallback(AnimationCallback callback)
 {
     _clock.setAnimationCallback(callback);
 }
 
-void Clock::Animation::removeAnimationCallback()
+void Animation::removeAnimationCallback()
 {
     _clock.setAnimationCallback(nullptr);
 }
 
-void Clock::Animation::setLoopCallback(LoopCallback callback)
+void Animation::setLoopCallback(LoopCallback callback)
 {
     _callback = callback;
 }
 
-void Clock::Animation::removeLoopCallback()
+void Animation::removeLoopCallback()
 {
     _callback = nullptr;
 }
@@ -286,12 +288,12 @@ void Clock::Animation::removeLoopCallback()
 // ------------------------------------------------------------------------
 // Fading Animation
 
-Clock::FadingAnimation::FadingAnimation(ClockPlugin &clock, Color from, Color to, float speed, uint16_t time, Color factor) : Animation(clock), _from(from), _to(to), _time(time), _speed(_secondsToSpeed(speed)), _factor(factor)
+FadingAnimation::FadingAnimation(ClockPlugin &clock, Color from, Color to, float speed, uint16_t time, Color factor) : Animation(clock), _from(from), _to(to), _time(time), _speed(_secondsToSpeed(speed)), _factor(factor)
 {
     __LDBG_printf("from=%s to=%s time=%u speed=%u (%f) factor=%s", _from.toString().c_str(), _to.toString().c_str(), _time, _speed, speed, _factor.toString().c_str());
 }
 
-void Clock::FadingAnimation::begin()
+void FadingAnimation::begin()
 {
     _finished = false;
     _progress = 0;
@@ -304,12 +306,12 @@ void Clock::FadingAnimation::begin()
     });
 }
 
-uint16_t Clock::FadingAnimation::_secondsToSpeed(float seconds) const
+uint16_t FadingAnimation::_secondsToSpeed(float seconds) const
 {
     return (kProgressPerSecond / seconds) + 0.5;
 }
 
-void Clock::FadingAnimation::callback(time_t now)
+void FadingAnimation::callback(time_t now)
 {
     if (_progress < kMaxProgress) {
         _progress = std::min((uint32_t)kMaxProgress, (uint32_t)(_progress + _speed));
@@ -354,7 +356,7 @@ void Clock::FadingAnimation::callback(time_t now)
 // ------------------------------------------------------------------------
 // Rainbow Animation
 
-Clock::RainbowAnimation::RainbowAnimation(ClockPlugin &clock, uint16_t speed, RainbowMultiplier &multiplier, RainbowColor &color) :
+RainbowAnimation::RainbowAnimation(ClockPlugin &clock, uint16_t speed, RainbowMultiplier &multiplier, RainbowColor &color) :
     Animation(clock),
     _speed(speed),
     _multiplier(multiplier),
@@ -364,7 +366,7 @@ Clock::RainbowAnimation::RainbowAnimation(ClockPlugin &clock, uint16_t speed, Ra
     // __LDBG_printf("speed=%u _multiplier=%f multiplier=%f factor=%s mins=%s", _speed, _multiplier, multiplier, _factor.toString().c_str(), _minimum.toString().c_str());
 }
 
-Clock::Color Clock::RainbowAnimation::_normalizeColor(uint8_t red, uint8_t green, uint8_t blue) const
+Color RainbowAnimation::_normalizeColor(uint8_t red, uint8_t green, uint8_t blue) const
 {
     return Color(
         std::max(red, _color.min.red),
@@ -373,8 +375,12 @@ Clock::Color Clock::RainbowAnimation::_normalizeColor(uint8_t red, uint8_t green
     );
 }
 
-float Clock::RainbowAnimation::_increment(float value, float min, float max, float incr)
+void RainbowAnimation::_increment(float *valuePtr, float min, float max, float *incrPtr)
 {
+    float value;
+    float incr;
+    memcpy(&value, valuePtr, sizeof(value));
+    memcpy(&incr, incrPtr, sizeof(incr));
     if (incr) {
         value += incr;
         if (value < min && incr < 0) {
@@ -382,14 +388,15 @@ float Clock::RainbowAnimation::_increment(float value, float min, float max, flo
             incr = -incr;
         }
         else if (value >= max && incr > 0) {
-            value = max - 0.00001;
             incr = -incr;
+            value = max + incr;
         }
     }
-    return value;
+    memcpy(valuePtr, &value, sizeof(*valuePtr));
+    memcpy(incrPtr, &incr, sizeof(*incrPtr));
 }
 
-void Clock::RainbowAnimation::begin()
+void RainbowAnimation::begin()
 {
     _finished = false;
     setAnimationCallback([this](PixelAddressType address, ColorType color, const SevenSegmentDisplay::Params_t &params) -> ColorType {
@@ -417,29 +424,29 @@ void Clock::RainbowAnimation::begin()
     __LDBG_printf("begin rate=%u", getUpdateRate());
 }
 
-void Clock::RainbowAnimation::end()
+void RainbowAnimation::end()
 {
     __LDBG_printf("end rate=%u", getUpdateRate());
     Animation::end();
 }
 
-void Clock::RainbowAnimation::loop(time_t now)
+void RainbowAnimation::loop(time_t now)
 {
-    _multiplier.value = _increment(_multiplier.value, _multiplier.min, _multiplier.max, _multiplier.incr);
-    _factor._red = _increment(_factor._red, _color.min.red, 256, _color.red_incr);
-    _factor._green = _increment(_factor._green, _color.min.green, 256, _color.green_incr);
-    _factor._blue = _increment(_factor._blue, _color.min.blue, 256, _color.blue_incr);
+    _increment(&_multiplier.value, _multiplier.min, _multiplier.max, &_multiplier.incr);
+    _increment(&_factor._red, _color.min.red, 256, &_color.red_incr);
+    _increment(&_factor._green, _color.min.green, 256, &_color.green_incr);
+    _increment(&_factor._blue, _color.min.blue, 256, &_color.blue_incr);
 }
 
 // ------------------------------------------------------------------------
 // Flashing Animation
 
-Clock::FlashingAnimation::FlashingAnimation(ClockPlugin &clock, Color color, uint16_t time, uint8_t mod) : Animation(clock), _color(color), _time(time), _mod(mod)
+FlashingAnimation::FlashingAnimation(ClockPlugin &clock, Color color, uint16_t time, uint8_t mod) : Animation(clock), _color(color), _time(time), _mod(mod)
 {
     __LDBG_printf("color=%s time=%u", _color.toString().c_str(), _time);
 }
 
-void Clock::FlashingAnimation::begin()
+void FlashingAnimation::begin()
 {
     __LDBG_printf("begin color=%s time=%u", _color.toString().c_str(), _time);
     _finished = false;
@@ -450,22 +457,53 @@ void Clock::FlashingAnimation::begin()
     setUpdateRate(std::max((uint16_t)(ClockPlugin::kMinFlashingSpeed / _mod), (uint16_t)(_time / _mod)));
 }
 
-void Clock::FlashingAnimation::end()
+void FlashingAnimation::end()
 {
     __LDBG_printf("end color=%s time=%u", _color.toString().c_str(), _time);
     Animation::end();
 }
 
 // ------------------------------------------------------------------------
-// Flashing Animation
+SkipRowsAnimation::SkipRowsAnimation(ClockPlugin &clock, uint16_t rows, uint16_t cols, uint32_t time) :
+    Animation(clock),
+    _rows(rows),
+    _cols(cols),
+    _time(time)
+{
+    setAnimationCallback([this](PixelAddressType address, ColorType color, const SevenSegmentDisplay::Params_t &params) -> ColorType {
+        uint16_t start = _time ? params.millis / _time : 0;
+        if (_rows > 1 && (address / IOT_LED_MATRIX_COLS / _rows) != (start % IOT_LED_MATRIX_ROWS)) {
+            return 0;
+        }
+        if (_cols > 1 && (address % IOT_LED_MATRIX_COLS / _cols) != (start % IOT_LED_MATRIX_COLS)) {
+            return 0;
+        }
+        return SevenSegmentDisplay::_adjustBrightnessAlways(color, params.brightness);
+    });
+    setUpdateRate(std::min<uint32_t>(std::max(5U, _time), ClockPlugin::kDefaultUpdateRate));
+}
 
-Clock::CallbackAnimation::CallbackAnimation(ClockPlugin &clock, AnimationCallback callback, uint16_t updateRate, bool blinkColon) :
+void SkipRowsAnimation::begin()
+{
+    __LDBG_printf("begin rows=%u cols=%u time=%u", _rows, _cols, _time);
+}
+
+void SkipRowsAnimation::end()
+{
+    __LDBG_printf("end rows=%u cols=%u time=%u", _rows, _cols, _time);
+    Animation::end();
+}
+
+// ------------------------------------------------------------------------
+// Callback Animation
+
+CallbackAnimation::CallbackAnimation(ClockPlugin &clock, AnimationCallback callback, uint16_t updateRate, bool blinkColon) :
     Animation(clock), _callback(callback), _updateRate(updateRate), _blinkColon(blinkColon)
 {
     __LDBG_printf("callback=%p/%u update_rate=%u blink_colon=%u", &_callback, (bool)_callback, _updateRate, _blinkColon);
 }
 
-void Clock::CallbackAnimation::begin()
+void CallbackAnimation::begin()
 {
     __LDBG_printf("begin");
     _finished = false;
@@ -474,7 +512,7 @@ void Clock::CallbackAnimation::begin()
     setAnimationCallback(_callback);
 }
 
-void Clock::CallbackAnimation::end()
+void CallbackAnimation::end()
 {
     __LDBG_printf("end");
     Animation::end();