+++ src/plugins/clock/clock_config.cpp
@@ -7,7 +7,6 @@
 
 namespace KFCConfigurationClasses {
 
-
     Plugins::ClockConfig::RainbowMultiplier_t::RainbowMultiplier_t() :
         value(1.23),
         min(0.1),
@@ -18,13 +17,13 @@ namespace KFCConfigurationClasses {
     Plugins::ClockConfig::RainbowColor_t::RainbowColor_t() :
         min(0x000000),
         factor(0xffffff),
-        red_incr(0.01),
-        green_incr(0.015),
-        blue_incr(0.012)
+        red_incr(0.1),
+        green_incr(0.15),
+        blue_incr(0.125)
     {}
 
     Plugins::ClockConfig::ClockConfig_t::ClockConfig_t() :
-        solid_color{ 0x00ff00 },
+        solid_color(0x00ff00),
         animation(0),
 #if !IOT_LED_MATRIX
         time_format_24h(true),
@@ -35,10 +34,14 @@ namespace KFCConfigurationClasses {
         blink_colon_speed(1000),
 #endif
         flashing_speed(150),
-        protection{ 55, 65, 75 },
+        protection( { { 55, 85 }, 75} ),
         rainbow{ RainbowMultiplier_t(), RainbowColor_t(), 30 },
-        alarm{{ 0xff0000 }, 250},
-        fading{ 7.5f, 2, { 0xffffff }}
+        alarm{ { 0xff0000 }, 250 },
+        fading{ .5f, 2, 0xffffff }
+#if IOT_LED_MATRIX
+        ,
+        skip_rows({ 2, 0, 60000 })
+#endif
     {
     }